Uploading first version of RektNet accessor
Uploading a python script which demonstrates how to use the RektNet dataset from MIT by selecting a random image from the dataset and overlay the cones landmark above it. Also uploading a yml file to recreate the conda environment used to run the script (NB: the yml file also contains pytorch which isn't required for the script to run).